fix performance issues with mouse tracking
Created Shape child component from Overlay that will handle mouse tracking
